+++ src/main/java/com/smoorbuilderserver/repository/CombatActionDescriptionRepository.java
@@ -1,5 +1,6 @@
 package com.smoorbuilderserver.repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
@@ -8,13 +9,13 @@ import org.springframework.stereotype.Repository;
 import com.smoorbuilderserver.model.CombatActionDescription;
 
 @Repository
-public interface CombatActionDescriptionRepository extends CrudRepository<CombatActionDescription, Long> {
+public interface CombatActionDescriptionRepository extends CrudRepository<CombatActionDescription, BigInteger> {
 	
 	public CombatActionDescription findByActionName(String actionName);
 	
 	public CombatActionDescription findByActionDescription(String actionDescription);
 	
-	public CombatActionDescription findByid(Integer id);
+	public CombatActionDescription findByid(BigInteger id);
 	
 	public List<CombatActionDescription> findDistinctCombatActionDescriptionByActionName(String actionName);
 }
