+++ src/main/java/com/smoorbuilderserver/repository/GeneralSkillDescriptionRepository.java
@@ -1,5 +1,6 @@
 package com.smoorbuilderserver.repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
@@ -8,11 +9,11 @@ import org.springframework.stereotype.Repository;
 import com.smoorbuilderserver.model.GeneralSkillDescription;
 
 @Repository
-public interface GeneralSkillDescriptionRepository extends CrudRepository<GeneralSkillDescription, Long> {
+public interface GeneralSkillDescriptionRepository extends CrudRepository<GeneralSkillDescription, BigInteger> {
 	
 	public GeneralSkillDescription findByGeneralSkillName(String generalSkillName);
 	
-	public GeneralSkillDescription findByid(Integer id);
+	public GeneralSkillDescription findByid(BigInteger id);
 	
 	public List<GeneralSkillDescription> findDistinctGeneralSkillDescriptionByGeneralSkillName(String generalSkillName);
 }
