+++ src/main/java/com/smoorbuilderserver/repository/ProductionSkillDescriptionRepository.java
@@ -1,5 +1,6 @@
 package com.smoorbuilderserver.repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
@@ -8,11 +9,11 @@ import org.springframework.stereotype.Repository;
 import com.smoorbuilderserver.model.ProductionSkillDescription;
 
 @Repository
-public interface ProductionSkillDescriptionRepository extends CrudRepository<ProductionSkillDescription, Long> {
+public interface ProductionSkillDescriptionRepository extends CrudRepository<ProductionSkillDescription, BigInteger> {
 	
 	public ProductionSkillDescription findByProductionSkillName(String productionlName);
 	
-	public ProductionSkillDescription findByid(Integer id);
+	public ProductionSkillDescription findByid(BigInteger id);
 	
 	public List<ProductionSkillDescription> findDistinctProductionSkillDescriptionByProductionSkillName(String productionName);
 }
