+++ src/main/java/com/smoorbuilderserver/repository/SpellcastingDescriptionRepository.java
@@ -1,5 +1,6 @@
 package com.smoorbuilderserver.repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
@@ -8,11 +9,11 @@ import org.springframework.stereotype.Repository;
 import com.smoorbuilderserver.model.SpellcastingDescription;
 
 @Repository
-public interface SpellcastingDescriptionRepository extends CrudRepository<SpellcastingDescription, Long> {
+public interface SpellcastingDescriptionRepository extends CrudRepository<SpellcastingDescription, BigInteger> {
 	
 	public SpellcastingDescription findBySpellcastingName(String generalSkillName);
 	
-	public SpellcastingDescription findByid(Integer id);
+	public SpellcastingDescription findByid(BigInteger id);
 	
 	public List<SpellcastingDescription> findDistinctSpellcastingDescriptionBySpellcastingName(String spellcastingDescriptionName);
 }
